Reorder connection markers in TAP tests
Commit 7f5b198 introduced TAP tests that use string literals to mark the
presence of a query in server logs. For no explicable reason, the tests
with the marker 'connect2' occur before the tests that use 'connect1'
marker.

So swap the connection marker strings so that a reader doesn't have to
spend extra effort in trying to decipher why 'connect2' tests appear
before 'connect1' tests.